refactor: tmp 지우기 및 genre 폴더링 변경
