+++ src/components/Ranking/RankingCard.tsx
