+++ src/components/Ranking/types.ts
