+++ src/components/battle/List/constants.ts
