+++ src/components/battle/List/types.ts
