+++ src/components/battle/List/useGetBattles.ts
