+++ src/components/battle/VoteResult/index.tsx
