+++ src/components/battle/VoteResult/useGetVoteResult.ts
