+++ src/components/battle/detail/Battle/index.tsx
@@ -10,7 +10,7 @@ import { COLOR } from '@/constants/color';
 import { useToast } from '@/hooks/useToast';
 
 import { SelectedBattle } from '../../types';
-import VoteResult from '../../VoteResult_tmp';
+import VoteResult from '../../VoteResult';
 import BattleMusic from '../BattleMusic';
 import { BattleMusicWrapper, Container } from './style';
