+++ src/components/post/Detail/index.tsx
