+++ src/components/post/Detail/types.ts
