+++ src/components/post/battle/selectMusic/index.tsx
@@ -4,7 +4,7 @@ import MusicPlusButton from 'public/images/plus-music.svg';
 import { StyledMusicPlayButton } from '@/components/battle/detail/BattleMusic/style';
 import { COLOR } from '@/constants/color';
 
-import { Music } from '../../Detail_tmp/types';
+import { Music } from '../../Detail/types';
 
 interface Prop {
   music: Music;
