+++ src/components/profile/UserContent.tsx
@@ -7,7 +7,7 @@ import FinishedBattleCard from '@/components/battle/Card/Finished';
 import NoContent from '@/components/common/NoContent';
 import RecommendationPost from '@/components/common/RecommendationPost';
 
-import { FinishedBattleMusic } from '../battle/List_tmp/types';
+import { FinishedBattleMusic } from '../battle/List/types';
 import { getPostFeedLimit } from './api';
 import { useGetMyBattleList } from './battle/useGetMyBattleList';
 import ContentList from './ContentList';
