+++ src/components/profile/battle/useGetMyBattleList.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { GenreValue } from '@/components/battle/List_tmp/types';
+import { GenreValue } from '@/components/battle/List/types';
 
 import { BattleStatusValue } from './../../battle/types';
 import { getMyBattleList } from './api';
