+++ src/pages/profile/battle/index.tsx
@@ -2,16 +2,16 @@ import styled from '@emotion/styled';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
 
+import AuthRequiredPage from '@/components/auth/AuthRequiredPage';
 import { BATTLE_STATUS_NAME_LIST, BATTLE_STATUS_VALUE_MAP } from '@/components/battle/constants';
-import BattleList from '@/components/battle/List_tmp';
-import { GenreValue, isGenreValue } from '@/components/battle/List_tmp/types';
+import BattleList from '@/components/battle/List';
+import { GenreValue, isGenreValue } from '@/components/battle/List/types';
 import { BattleStatusName } from '@/components/battle/types';
 import BottomNav from '@/components/common/BottomNav';
 import Filter from '@/components/common/Filter';
 import Genres from '@/components/common/Genres';
 import Header from '@/components/common/Header';
 import NoContent from '@/components/common/NoContent';
-import AuthRequiredPage from '@/components/auth/AuthRequiredPage';
 import { useGetMyBattleList } from '@/components/profile/battle/useGetMyBattleList';
 
 export default function MyBattleListPage() {
